+++ app/src/main/java/com/example/productivitycac/LoginActivity.java
@@ -4,7 +4,10 @@ import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.Button;
 import android.widget.EditText;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.snackbar.Snackbar;
@@ -16,6 +19,56 @@ public class LoginActivity  extends AppCompatActivity {
 
     EditText username = (EditText) findViewById(R.id.editText1);
     EditText password = (EditText) findViewById(R.id.editText2);
+
+
+    Button b1,b2;
+    EditText ed1,ed2;
+
+    TextView tx1;
+    int counter = 3;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        b1 = (Button)findViewById(R.id.button);
+        ed1 = (EditText)findViewById(R.id.editText);
+        ed2 = (EditText)findViewById(R.id.editText2);
+
+        b2 = (Button)findViewById(R.id.button2);
+        // tx1 = (TextView)findViewById(R.id.textView3);
+        // tx1.setVisibility(View.GONE);
+
+        b1.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if(ed1.getText().toString().equals("admin") &&
+                        ed2.getText().toString().equals("admin")) {
+                    Toast.makeText(getApplicationContext(),
+                            "Redirecting...",Toast.LENGTH_SHORT).show();
+                }else{
+                    Toast.makeText(getApplicationContext(), "Wrong Credentials",Toast.LENGTH_SHORT).show();
+
+                    //        tx1.setVisibility(View.VISIBLE);
+                    // tx1.setBackgroundColor(Color.RED);
+                    counter--;
+                    // tx1.setText(Integer.toString(counter));
+
+                    if (counter == 0) {
+                        b1.setEnabled(false);
+                    }
+                }
+            }
+        });
+
+        b2.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                finish();
+            }
+        });
+    }
 
     public void login(View view) {
         if (username.getText().toString().equals("admin") && password.getText().toString().equals("admin")) {
